assorted bugfixes
git-svn-id: https://gnunet.org/svn/GNUnet@1571 140774ce-b5e7-0310-ab8b-a85725594a96

+++ src/applications/fs/ecrs/download.c
@@ -862,8 +862,8 @@ static int nodeReceive(const HashCode512 * query,
   size = ntohl(reply->size) - sizeof(Datastore_Value);
   if ( (size <= sizeof(DBlock)) ||
        (size - sizeof(DBlock) != getNodeSize(node)) ) {
-    printf("Received %u bytes, expected %u\n",
-	   size - sizeof(DBlock),
+    printf("Received %llu bytes, expected %u\n",
+	   (unsigned long long) (size - sizeof(DBlock)),
 	   getNodeSize(node));
     BREAK();
     return SYSERR; /* invalid size! */
